+++ src/actions.cpp
@@ -1220,7 +1220,7 @@ namespace Action
     int Insert::insertXmpPacket(const std::string& path, const Exiv2::DataBuf& xmpBlob, bool usePacket) const
     {
         std::string xmpPacket;
-        for (long i = 0; i < xmpBlob.size_; i++) {
+        for (size_t i = 0; i < xmpBlob.size_; i++) {
             xmpPacket += static_cast<char>(xmpBlob.pData_[i]);
         }
         Exiv2::Image::UniquePtr image = Exiv2::ImageFactory::open(path);
