+++ src/pngchunk_int.cpp
@@ -109,12 +109,12 @@ namespace Exiv2 {
     {
         // From a tEXt, zTXt, or iTXt chunk,
         // we get the key, it's a null terminated string at the chunk start
-        const int offset = stripHeader ? 8 : 0;
+        const size_t offset = stripHeader ? 8 : 0;
         if (data.size_ <= offset) throw Error(kerFailedToReadImageData);
         const byte *key = data.pData_ + offset;
 
         // Find null string at end of key.
-        int keysize=0;
+        size_t keysize=0;
         while (key[keysize] != 0)
         {
             keysize++;
@@ -128,14 +128,14 @@ namespace Exiv2 {
     } // PngChunk::keyTXTChunk
 
     DataBuf PngChunk::parseTXTChunk(const DataBuf& data,
-                                    int            keysize,
+                                    size_t        keysize,
                                     TxtChunkType   type)
     {
         DataBuf arr;
 
         if(type == zTXt_Chunk)
         {
-            enforce(data.size_ >= Safe::add(keysize, 2), Exiv2::kerCorruptedMetadata);
+            enforce(data.size_ >= Safe::add(keysize, 2ul), Exiv2::kerCorruptedMetadata);
 
             // Extract a deflate compressed Latin-1 text chunk
 
@@ -152,14 +152,14 @@ namespace Exiv2 {
 
             // compressed string after the compression technique spec
             const byte* compressedText      = data.pData_ + keysize + 2;
-            long compressedTextSize = data.size_  - keysize - 2;
+            size_t compressedTextSize = data.size_  - keysize - 2;
             enforce(compressedTextSize < data.size_, kerCorruptedMetadata);
 
             zlibUncompress(compressedText, compressedTextSize, arr);
         }
         else if(type == tEXt_Chunk)
         {
-            enforce(data.size_ >= Safe::add(keysize, 1), Exiv2::kerCorruptedMetadata);
+            enforce(data.size_ >= Safe::add(keysize, 1ul), Exiv2::kerCorruptedMetadata);
             // Extract a non-compressed Latin-1 text chunk
 
             // the text comes after the key, but isn't null terminated
@@ -170,7 +170,7 @@ namespace Exiv2 {
         }
         else if(type == iTXt_Chunk)
         {
-            enforce(data.size_ >= Safe::add(keysize, 3), Exiv2::kerCorruptedMetadata);
+            enforce(data.size_ >= Safe::add(keysize, 3ul), Exiv2::kerCorruptedMetadata);
             const size_t nullSeparators = std::count(&data.pData_[keysize+3], &data.pData_[data.size_], '\0');
             enforce(nullSeparators >= 2, Exiv2::kerCorruptedMetadata);
 
@@ -187,11 +187,11 @@ namespace Exiv2 {
             // language description string after the compression technique spec
             const size_t languageTextMaxSize = data.size_ - keysize - 3;
             std::string languageText =
-                string_from_unterminated((const char*)(data.pData_ + Safe::add(keysize, 3)), languageTextMaxSize);
+                string_from_unterminated((const char*)(data.pData_ + Safe::add(keysize, 3ul)), languageTextMaxSize);
             const size_t languageTextSize = languageText.size();
 
             enforce(static_cast<unsigned long>(data.size_) >=
-                    Safe::add(static_cast<size_t>(Safe::add(keysize, 4)), languageTextSize),
+                    Safe::add(static_cast<size_t>(Safe::add(keysize, 4ul)), languageTextSize),
                     Exiv2::kerCorruptedMetadata);
             // translated keyword string after the language description
             std::string translatedKeyText =
@@ -650,7 +650,7 @@ namespace Exiv2 {
         }
         sp++ ; // step over '\n'
 
-        long length = (long) atol(startOfLength);
+        size_t length = (size_t) atol(startOfLength);
 
         // Allocate space
         if (length == 0)
